+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- support `(impl ToTokens, impl ToTokens)` tuples for span range
+
 ## [0.9.0] - 2023-11-06
 ### Added
 - support `impl Parse` inputs and `impl ToTokens` outputs.